backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Mail==0.9.1
python-dotenv==1.0.0
cloudinary==1.36.0
sendgrid==6.11.0
requests==2.31.0
pytest==7.4.3
pytest-flask==1.3.0
flasgger==0.9.7.1